requirements.txt
pymem
PyQt5
psutil
capstone
keystone-engine
PyYAML
pywin32